pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.10.0",
    "groq>=0.18.0",
    "langchain>=0.3.18",
    "numpy>=1.26.4",
    "openai>=1.63.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.0",
]
